Move arangosearch-memory-metrics tests into separate files for better distribution into test buckets.
