+++ tests/js/client/shell/arangosearch-memory-metrics-cluster.js
@@ -27,36 +27,22 @@
 const jsunity = require("jsunity");
 const internal = require("internal");
 const deriveTestSuite = require('@arangodb/test-helper').deriveTestSuite;
-const base = require("fs").join(internal.pathForTesting('client'), 
-    'shell', 'shell-improved-metrics-accounting.inc');
-const ImprovedMemoryAccounting = internal.load(base);
 
-////////////////////////////////////////////////////////////////////////////////
-/// @brief executes the test suite
-////////////////////////////////////////////////////////////////////////////////
+const arangosearch_base = require("fs").join(internal.pathForTesting('client'), 
+    'shell', 'api', 'arangosearch-memory-metrics.inc');
+const ImprovedMemoryAccountingArangoSearch = internal.load(arangosearch_base);
 
-if (internal.debugCanUseFailAt()) {
-    const base_fail_at = require("fs").join(internal.pathForTesting('client'), 
-    'shell', 'shell-improved-metrics-accounting-fail-at.inc');
-    const ImprovedMemoryAccountingFailAt = internal.load(base_fail_at);
-
-    jsunity.run(function ImprovedMemoryAccountingFailAtTestSuite_no_repl() {
-        let suite = {};
-        deriveTestSuite(
-            ImprovedMemoryAccountingFailAt("ImprovedMemoryAccountingFailAtTestSuite_NoRepl",null, {}),
-            suite,
-            "_NoRepl"
-        );
-        return suite;
-      });
-}
-
-jsunity.run(function ImprovedMemoryAccountingTestSuite_no_repl() {
-    let suite = {};  
+jsunity.run(function ImprovedMemoryAccountingArangoSearchTestSuite_repl() {
+    let suite = {
+    };
+  
     deriveTestSuite(
-      ImprovedMemoryAccounting("ImprovedMemoryAccountingTestSuite_NoRepl", null, {}),
+      ImprovedMemoryAccountingArangoSearch("ImprovedMemoryAccountingArangoSearch_Repl", null, {
+        replicationFactor:3, 
+        writeConcern:1, 
+        numberOfShards : 3}, false),
         suite,
-        "_NoRepl"
+        "_Repl"
     );
     return suite;
 });
