+++ tests/js/client/shell/shell-improved-metrics-accounting-oneshard-cluster.js
@@ -27,28 +27,21 @@
 const jsunity = require("jsunity");
 const internal = require("internal");
 const deriveTestSuite = require('@arangodb/test-helper').deriveTestSuite;
-const base = require("fs").join(require('internal').pathForTesting('client'), 
+const base = require("fs").join(internal.pathForTesting('client'), 
     'shell', 'shell-improved-metrics-accounting.inc');
-
-const arangosearch_base = require("fs").join(require('internal').pathForTesting('client'), 
-    'shell', 'api', 'arangosearch-memory-metrics.inc');    
-
-const ImprovedMemoryAccounting = require("internal").load(base);
-const ImprovedMemoryAccountingArangoSearch = require("internal").load(arangosearch_base);
+const ImprovedMemoryAccounting = internal.load(base);
 
 ////////////////////////////////////////////////////////////////////////////////
 /// @brief executes the test suite
 ////////////////////////////////////////////////////////////////////////////////
 
 if (internal.debugCanUseFailAt()) {
-    const base_fail_at = require("fs").join(require('internal').pathForTesting('client'), 
+    const base_fail_at = require("fs").join(internal.pathForTesting('client'), 
     'shell', 'shell-improved-metrics-accounting-fail-at.inc');
-    const ImprovedMemoryAccountingFailAt = require("internal").load(base_fail_at);
+    const ImprovedMemoryAccountingFailAt = internal.load(base_fail_at);
 
     jsunity.run(function ImprovedMemoryAccountingFailAtTestSuite_oneshard() {
-        let suite = {
-        };
-      
+        let suite = {};
         deriveTestSuite(
             ImprovedMemoryAccountingFailAt("InvertedIndexSearchAliasqlTestSuite_OneShard", {sharding: "single"}, {}),
             suite,
@@ -59,9 +52,7 @@ if (internal.debugCanUseFailAt()) {
 }
 
 jsunity.run(function ImprovedMemoryAccountingTestSuite_oneshard() {
-    let suite = {
-    };
-  
+    let suite = {};
     deriveTestSuite(
       ImprovedMemoryAccounting("ImprovedMemoryAccountingTestSuite_OneShard", {sharding: "single"}, {}),
         suite,
@@ -70,15 +61,4 @@ jsunity.run(function ImprovedMemoryAccountingTestSuite_oneshard() {
     return suite;
 });
 
-jsunity.run(function ImprovedMemoryAccountingArangoSearchTestSuite_oneshard() {
-    let suite = {
-    };
-  
-    deriveTestSuite(
-      ImprovedMemoryAccountingArangoSearch("ImprovedMemoryAccountingArangoSearch_OneShard", {sharding: "single"}, {}),
-        suite,
-        "_OneShard"
-    );
-    return suite;
-});
 return jsunity.done();
